pyproject.toml
[project]
name = "predict-mlb"
version = "0.1.0"
description = "MLB 선수 기록 조회 및 예측 서비스"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair==5.5.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.0",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "idna==3.10",
    "jinja2==3.1.6",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "markupsafe==3.0.2",
    "narwhals==1.39.1",
    "numpy==2.2.6",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "protobuf==6.31.0",
    "pyarrow==20.0.0",
    "pydeck==0.9.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rpds-py==0.25.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "streamlit==1.45.1",
    "streamlit-option-menu==0.4.0",
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tornado==6.5",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "watchdog==6.0.0 ; sys_platform != 'darwin'",
    "prophet>=1.1.4",
    "matplotlib>=3.7.1",
    "seaborn>=0.12.2",
    "streamlit-option-menu>=0.3.2",
]

[project.urls]
Homepage = "https://github.com/yourusername/predict_mlb"
Issues = "https://github.com/yourusername/predict_mlb/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "logs/",
    "__pycache__/",
    "*.pyc",
    "*.pyo",
    "predict_mlb/",
    ".git/",
    ".venv/",
    "tests/"
]

[tool.hatch.metadata]
allow-direct-references = true
